frameworks/docling/requirements.txt
docling
pandas
python-dotenv
